Only delete preferences on API>=24

+++ project/app/src/main/java/org/owntracks/android/preferences/SharedPreferencesStore.kt
@@ -2,6 +2,7 @@ package org.owntracks.android.preferences
 
 import android.content.Context
 import android.content.SharedPreferences
+import android.os.Build
 import androidx.preference.PreferenceManager
 import dagger.hilt.android.qualifiers.ApplicationContext
 import javax.inject.Inject
@@ -60,9 +61,11 @@ class SharedPreferencesStore @Inject constructor(@ApplicationContext private val
                     }
                 }
                 oldSharedPreferenceNames.forEach {
-                    val deleted = context.deleteSharedPreferences(it)
-                    if (!deleted) {
-                        Timber.e("Failed to delete shared preference $it")
+                    if (android.os.Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
+                        val deleted = context.deleteSharedPreferences(it)
+                        if (!deleted) {
+                            Timber.e("Failed to delete shared preference $it")
+                        }
                     }
                 }
             }
